And-EndOfDayThree

+++ Assets/Scripts/Interfaces/IHealth.cs
@@ -4,6 +4,6 @@ public interface IHealth
 {
     float CurrentHealth { get; }
     float MaxHealth { get; }
-    void Heal(int amount); 
+    void Heal(float amount); 
     void Die();
 }
